大本で padding を確保するのをやめる。余ったスペースでお絵かき
