+++ app/posts/tag/[tag]/[page]/page.tsx
@@ -41,7 +41,7 @@ export default async function TagListPage(props: PageProps) {
 
     return (
         <>
-            <div className="max-w-6xl m-auto flex flex-col space-y-4">
+            <div className="p-4 max-w-6xl m-auto flex flex-col space-y-4">
 
                 <Title
                     title={unEscapeText}
