+++ app/posts/tag/all_tags/page.tsx
@@ -15,7 +15,7 @@ export default async function ClientAllTagPage() {
     const tagDataList = await ContentFolderManager.getAllTagDataList()
 
     return (
-        <div className="max-w-6xl m-auto flex flex-col space-y-4">
+        <div className="p-4 max-w-6xl m-auto flex flex-col space-y-4">
 
             <Title
                 title="タグ一覧"
